LayoutTests/http/tests/site-isolation/history/fragment-navigation-in-iframe-back.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true UsesBackForwardCache=false ] -->
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Verifies that fragment navigations in a cross-origin iframe are preserved across back navigation.");
jsTestIsAsync = true;

var reportCount = 0;
var iframeUrl = "";

onmessage = (event) => {
    iframeUrl = event.data;
    reportCount++;

    if (sessionStorage.verifyPhase) {
        // After back, iframe should be restored to the URL with fragment.
        shouldBeTrue("iframeUrl.endsWith('#section2')");
        delete sessionStorage.verifyPhase;
        finishJSTest();
        return;
    }

    if (reportCount === 1) {
        // Initial load complete. Tell iframe to navigate to #section2.
        document.getElementById("frame").contentWindow.postMessage("navigate-hash:section2", "*");
    } else if (reportCount === 2) {
        // hashchange reported. Navigate main frame away.
        sessionStorage.verifyPhase = "true";
        location.href = "/site-isolation/history/resources/go-back.html";
    }
};
</script>
<iframe id="frame" src="http://localhost:8000/site-isolation/history/resources/report-url-on-request.html"></iframe>
