LayoutTests/http/tests/site-isolation/history/fragment-navigation-in-iframe-back-expected.txt

Verifies that fragment navigations in a cross-origin iframe are preserved across back navigation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframeUrl.endsWith('#section2') is true
PASS successfullyParsed is true

TEST COMPLETE
